Move existing mosaic list into separate file
