+++ app/assets/scripts/components/project/prime-panel/tabs/predict/mosaic-selector/modal.js
@@ -5,10 +5,7 @@ import { Modal } from '@devseed-ui/modal';
 import { glsp } from '@devseed-ui/theme-provider';
 import { Button } from '@devseed-ui/button';
 import { Heading } from '@devseed-ui/typography';
-import CardList, { Card } from '../../../../../common/card-list';
-import { ProjectMachineContext } from '../../../../../../fsm/project';
-import selectors from '../../../../../../fsm/project/selectors';
-import { formatDateTime } from '../../../../../../utils/format';
+import { ExistingMosaicsSection } from './sections/list-mosaics';
 
 const ModalHeader = styled.header`
   padding: ${glsp(2)} ${glsp(2)} 0;
@@ -34,27 +31,7 @@ const Headline = styled.div`
   }
 `;
 
-const HeadingWrapper = styled.div`
-  display: flex;
-  flex-flow: row nowrap;
-  justify-content: space-between;
-  align-items: baseline;
-`;
-
-export function MosaicSelectorModal({ showModal, setShowModal, isProjectNew }) {
-  const actorRef = ProjectMachineContext.useActorRef();
-  const mosaicsList = ProjectMachineContext.useSelector(selectors.mosaicsList);
-  const currentMosaic = ProjectMachineContext.useSelector(
-    selectors.currentMosaic
-  );
-  const currentImagerySource = ProjectMachineContext.useSelector(
-    selectors.currentImagerySource
-  );
-
-  const selectableMosaics = mosaicsList.filter(
-    (mosaic) => mosaic.imagery_source_id === currentImagerySource?.id
-  );
-
+export function MosaicSelectorModal({ showModal, setShowModal }) {
   return (
     <Modal
       id='select-mosaic-modal'
@@ -66,8 +43,7 @@ export function MosaicSelectorModal({ showModal, setShowModal, isProjectNew }) {
       renderHeader={() => (
         <ModalHeader>
           <Headline>
-            {' '}
-            <Heading>Select an base mosaic</Heading>
+            <Heading>Select a base mosaic</Heading>
             <Button
               hideText
               variation='base-plain'
@@ -84,54 +60,7 @@ export function MosaicSelectorModal({ showModal, setShowModal, isProjectNew }) {
       )}
       content={
         <ModalContent>
-          <HeadingWrapper>
-            <Heading size='small' as='h4'>
-              Available mosaics for the AOI
-            </Heading>
-          </HeadingWrapper>
-          <CardList
-            nonScrolling
-            numColumns={2}
-            data={selectableMosaics}
-            renderCard={(mosaic) => {
-              const { name, mosaic_ts_end, mosaic_ts_start } = mosaic;
-
-              return (
-                <Card
-                  data-cy={`select-mosaic-${mosaic.id}-card`}
-                  key={mosaic.id}
-                  title={mosaic.name}
-                  details={{
-                    name,
-                    'Mosaic Start Date': mosaic_ts_start
-                      ? formatDateTime(mosaic_ts_start)
-                      : 'N/A',
-                    'Mosaic End Date': mosaic_ts_end
-                      ? formatDateTime(mosaic_ts_end)
-                      : 'N/A',
-                  }}
-                  borderlessMedia
-                  selected={currentMosaic && currentMosaic.id === mosaic.id}
-                  onClick={() => {
-                    if (isProjectNew) {
-                      if (!currentMosaic || currentMosaic.id !== mosaic.id) {
-                        actorRef.send({
-                          type: 'Mosaic was selected',
-                          data: { mosaic },
-                        });
-                      }
-                    } else if (currentMosaic?.id !== mosaic.id) {
-                      actorRef.send({
-                        type: 'Mosaic was selected',
-                        data: { mosaic },
-                      });
-                    }
-                    setShowModal(false);
-                  }}
-                />
-              );
-            }}
-          />
+          <ExistingMosaicsSection setShowModal={setShowModal} />
         </ModalContent>
       }
     />
@@ -139,7 +68,6 @@ export function MosaicSelectorModal({ showModal, setShowModal, isProjectNew }) {
 }
 
 MosaicSelectorModal.propTypes = {
-  isProjectNew: T.bool,
   showModal: T.bool,
   setShowModal: T.func.isRequired,
 };
